fix: break circular imports by extracting fns into a new individual_power.py
